chore: move net-config tests to new suite
This mirrors the change in file name prior.
